Simplify createTestUser helper and reorder test helpers
This does two things:

- Simplify `createTestUser` to not go through the whole `db.User.Create`
  flow, but instead insert two rows into the database directly.
- Move all test helpers and setup into `main_test.go`
