+++ enterprise/internal/campaigns/integration_test.go
@@ -1,16 +1,11 @@
 package campaigns
 
 import (
-	"database/sql"
-	"flag"
-	"strings"
 	"testing"
 
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbtesting"
 )
-
-var dsn = flag.String("dsn", "", "Database connection string to use in integration tests")
 
 func TestIntegration(t *testing.T) {
 	if testing.Short() {
@@ -21,7 +16,7 @@ func TestIntegration(t *testing.T) {
 
 	dbtesting.SetupGlobalTestDB(t)
 
-	userID := insertTestUser(t, dbconn.Global)
+	user := createTestUser(t, false)
 
 	t.Run("Store", func(t *testing.T) {
 		t.Run("Campaigns", storeTest(dbconn.Global, testStoreCampaigns))
@@ -33,38 +28,7 @@ func TestIntegration(t *testing.T) {
 		t.Run("CodeHosts", storeTest(dbconn.Global, testStoreCodeHost))
 	})
 
-	t.Run("GitHubWebhook", testGitHubWebhook(dbconn.Global, userID))
-	t.Run("BitbucketWebhook", testBitbucketWebhook(dbconn.Global, userID))
-	t.Run("GitLabWebhook", testGitLabWebhook(dbconn.Global, userID))
-}
-
-func truncateTables(t *testing.T, db *sql.DB, tables ...string) {
-	t.Helper()
-
-	_, err := db.Exec("TRUNCATE " + strings.Join(tables, ", ") + " RESTART IDENTITY")
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func insertTestOrg(t *testing.T, db *sql.DB) (orgID int32) {
-	t.Helper()
-
-	err := db.QueryRow("INSERT INTO orgs (name) VALUES ('bbs-org') RETURNING id").Scan(&orgID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return orgID
-}
-
-func insertTestUser(t *testing.T, db *sql.DB) (userID int32) {
-	t.Helper()
-
-	err := db.QueryRow("INSERT INTO users (username) VALUES ('bbs-admin') RETURNING id").Scan(&userID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return userID
+	t.Run("GitHubWebhook", testGitHubWebhook(dbconn.Global, user.ID))
+	t.Run("BitbucketWebhook", testBitbucketWebhook(dbconn.Global, user.ID))
+	t.Run("GitLabWebhook", testGitLabWebhook(dbconn.Global, user.ID))
 }
